--- 2Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/keyword/repository/KeywordRepository.java
@@ -1,10 +1,0 @@
-package com.example.mutbooks.app.keyword.repository;
-
-import com.example.mutbooks.app.keyword.entity.PostKeyword;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface KeywordRepository extends JpaRepository<PostKeyword, Long> {
-    Optional<PostKeyword> findByContent(String content);
-}
