+++ 2Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/postKeyword/entity/PostKeyword.java
@@ -1,4 +1,4 @@
-package com.example.mutbooks.app.keyword.entity;
+package com.example.mutbooks.app.postKeyword.entity;
 
 import com.example.mutbooks.app.base.entity.BaseEntity;
 import lombok.Getter;
