+++ 2Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/postKeyword/repository/PostKeywordRepository.java
@@ -1,0 +1,10 @@
+package com.example.mutbooks.app.postKeyword.repository;
+
+import com.example.mutbooks.app.postKeyword.entity.PostKeyword;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PostKeywordRepository extends JpaRepository<PostKeyword, Long> {
+    Optional<PostKeyword> findByContent(String content);
+}
